refactor: merge config as params for AST transformer
